flaskpeewee/requirements.txt
flask
werkzeug
jinja2
wtforms
flask-peewee
gunicorn
gevent
